+++ df-chat-enhance/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Alpha Release 0.2.0
 - Added option for hiding the Export Chat Log button.
 - Added option to convert the other chat buttons into the same style as the new Roll Type buttons.
+- Moved the Manage Chat Archive button into a group so it will not conflict with any other buttons that get added to the Game Settings section of the settings tab.
 
 ## Alpha Release 0.1.0
 - First Alpha release with the core functionality of Chat Archive and Roll Selector Buttons.